requirements.txt
smart_open>=1.9.0
spacy>=2.2.3
vrt-generator>=0.0.6